Move interpolation functionality in calc to interpolate
This commit moves the interpolation functionality in metpy.calc.tools
(all forms of interpolation along a single axis) to a new module in
interpolate.

This also contains a series of changes to fix syntax errors and other
mistakes based on CI and code review
