+++ metpy/calc/tests/test_calc_tools.py
@@ -9,6 +9,7 @@ import numpy as np
 import numpy.ma as ma
 import pytest
 
+
 from metpy.calc import (find_bounding_indices, find_intersections, first_derivative, get_layer,
                         get_layer_heights, gradient, interp, interpolate_nans, laplacian,
                         log_interp, nearest_intersection_idx, parse_angle,
@@ -93,31 +94,14 @@ def test_find_intersections_intersections_in_data_at_ends(direction, expected):
 
 
 def test_interpolate_nan_linear():
-    """Test linear interpolation of arrays with NaNs in the y-coordinate."""
+    """Test deprecated interpolate_nans function."""
     x = np.linspace(0, 20, 15)
     y = 5 * x + 3
     nan_indexes = [1, 5, 11, 12]
     y_with_nan = y.copy()
     y_with_nan[nan_indexes] = np.nan
-    assert_array_almost_equal(y, interpolate_nans(x, y_with_nan), 2)
-
-
-def test_interpolate_nan_log():
-    """Test log interpolation of arrays with NaNs in the y-coordinate."""
-    x = np.logspace(1, 5, 15)
-    y = 5 * np.log(x) + 3
-    nan_indexes = [1, 5, 11, 12]
-    y_with_nan = y.copy()
-    y_with_nan[nan_indexes] = np.nan
-    assert_array_almost_equal(y, interpolate_nans(x, y_with_nan, kind='log'), 2)
-
-
-def test_interpolate_nan_invalid():
-    """Test log interpolation with invalid parameter."""
-    x = np.logspace(1, 5, 15)
-    y = 5 * np.log(x) + 3
-    with pytest.raises(ValueError):
-        interpolate_nans(x, y, kind='loog')
+    with pytest.warns(MetpyDeprecationWarning):
+        assert_array_almost_equal(y, interpolate_nans(x, y_with_nan), 2)
 
 
 @pytest.mark.parametrize('mask, expected_idx, expected_element', [
@@ -194,23 +178,25 @@ def test_delete_masked_points():
     assert_array_equal(b, expected)
 
 
+def test_interp():
+    """Test deprecated interp function."""
+    x = np.array([1., 2., 3., 4.])
+    y = x
+    x_interp = np.array([3.5000000, 2.5000000])
+    y_interp_truth = np.array([3.5000000, 2.5000000])
+    with pytest.warns(MetpyDeprecationWarning):
+        y_interp = interp(x_interp, x, y)
+    assert_array_almost_equal(y_interp, y_interp_truth, 7)
+
+
 def test_log_interp():
-    """Test interpolating with log x-scale."""
+    """Test deprecated log_interp function."""
     x_log = np.array([1e3, 1e4, 1e5, 1e6])
     y_log = np.log(x_log) * 2 + 3
     x_interp = np.array([5e3, 5e4, 5e5])
     y_interp_truth = np.array([20.0343863828, 24.6395565688, 29.2447267548])
-    y_interp = log_interp(x_interp, x_log, y_log)
-    assert_array_almost_equal(y_interp, y_interp_truth, 7)
-
-
-def test_log_interp_units():
-    """Test interpolating with log x-scale with units."""
-    x_log = np.array([1e3, 1e4, 1e5, 1e6]) * units.hPa
-    y_log = (np.log(x_log.m) * 2 + 3) * units.degC
-    x_interp = np.array([5e5, 5e6, 5e7]) * units.Pa
-    y_interp_truth = np.array([20.0343863828, 24.6395565688, 29.2447267548]) * units.degC
-    y_interp = log_interp(x_interp, x_log, y_log)
+    with pytest.warns(MetpyDeprecationWarning):
+        y_interp = log_interp(x_interp, x_log, y_log)
     assert_array_almost_equal(y_interp, y_interp_truth, 7)
 
 
@@ -352,131 +338,6 @@ def test_get_layer(pressure, variable, heights, bottom, depth, interp, expected)
                                  depth=depth, interpolate=interp)
     assert_array_almost_equal(p_layer, expected[0], 5)
     assert_array_almost_equal(y_layer, expected[1], 5)
-
-
-def test_log_interp_2d():
-    """Test interpolating with log x-scale in 2 dimensions."""
-    x_log = np.array([[1e3, 1e4, 1e5, 1e6], [1e3, 1e4, 1e5, 1e6]])
-    y_log = np.log(x_log) * 2 + 3
-    x_interp = np.array([5e3, 5e4, 5e5])
-    y_interp_truth = np.array([20.0343863828, 24.6395565688, 29.2447267548])
-    y_interp = log_interp(x_interp, x_log, y_log, axis=1)
-    assert_array_almost_equal(y_interp[1], y_interp_truth, 7)
-
-
-def test_log_interp_3d():
-    """Test interpolating with log x-scale 3 dimensions along second axis."""
-    x_log = np.ones((3, 4, 3)) * np.array([1e3, 1e4, 1e5, 1e6]).reshape(-1, 1)
-    y_log = np.log(x_log) * 2 + 3
-    x_interp = np.array([5e3, 5e4, 5e5])
-    y_interp_truth = np.array([20.0343863828, 24.6395565688, 29.2447267548])
-    y_interp = log_interp(x_interp, x_log, y_log, axis=1)
-    assert_array_almost_equal(y_interp[0, :, 0], y_interp_truth, 7)
-
-
-def test_log_interp_4d():
-    """Test interpolating with log x-scale 4 dimensions."""
-    x_log = np.ones((2, 2, 3, 4)) * np.array([1e3, 1e4, 1e5, 1e6])
-    y_log = np.log(x_log) * 2 + 3
-    x_interp = np.array([5e3, 5e4, 5e5])
-    y_interp_truth = np.array([20.0343863828, 24.6395565688, 29.2447267548])
-    y_interp = log_interp(x_interp, x_log, y_log, axis=3)
-    assert_array_almost_equal(y_interp[0, 0, 0, :], y_interp_truth, 7)
-
-
-def test_log_interp_2args():
-    """Test interpolating with log x-scale with 2 arguments."""
-    x_log = np.array([1e3, 1e4, 1e5, 1e6])
-    y_log = np.log(x_log) * 2 + 3
-    y_log2 = np.log(x_log) * 2 + 3
-    x_interp = np.array([5e3, 5e4, 5e5])
-    y_interp_truth = np.array([20.0343863828, 24.6395565688, 29.2447267548])
-    y_interp = log_interp(x_interp, x_log, y_log, y_log2)
-    assert_array_almost_equal(y_interp[1], y_interp_truth, 7)
-    assert_array_almost_equal(y_interp[0], y_interp_truth, 7)
-
-
-def test_log_interp_set_nan_above():
-    """Test interpolating with log x-scale setting out of bounds above data to nan."""
-    x_log = np.array([1e3, 1e4, 1e5, 1e6])
-    y_log = np.log(x_log) * 2 + 3
-    x_interp = np.array([1e7])
-    y_interp_truth = np.nan
-    with pytest.warns(Warning):
-        y_interp = log_interp(x_interp, x_log, y_log)
-    assert_array_almost_equal(y_interp, y_interp_truth, 7)
-
-
-def test_log_interp_no_extrap():
-    """Test interpolating with log x-scale setting out of bounds value error."""
-    x_log = np.array([1e3, 1e4, 1e5, 1e6])
-    y_log = np.log(x_log) * 2 + 3
-    x_interp = np.array([1e7])
-    with pytest.raises(ValueError):
-        log_interp(x_interp, x_log, y_log, fill_value=None)
-
-
-def test_log_interp_set_nan_below():
-    """Test interpolating with log x-scale setting out of bounds below data to nan."""
-    x_log = np.array([1e3, 1e4, 1e5, 1e6])
-    y_log = np.log(x_log) * 2 + 3
-    x_interp = 1e2
-    y_interp_truth = np.nan
-    with pytest.warns(Warning):
-        y_interp = log_interp(x_interp, x_log, y_log)
-    assert_array_almost_equal(y_interp, y_interp_truth, 7)
-
-
-def test_interp_2args():
-    """Test interpolation with 2 arguments."""
-    x = np.array([1., 2., 3., 4.])
-    y = x
-    y2 = x
-    x_interp = np.array([2.5000000, 3.5000000])
-    y_interp_truth = np.array([2.5000000, 3.5000000])
-    y_interp = interp(x_interp, x, y, y2)
-    assert_array_almost_equal(y_interp[0], y_interp_truth, 7)
-    assert_array_almost_equal(y_interp[1], y_interp_truth, 7)
-
-
-def test_interp_decrease():
-    """Test interpolation with decreasing interpolation points."""
-    x = np.array([1., 2., 3., 4.])
-    y = x
-    x_interp = np.array([3.5000000, 2.5000000])
-    y_interp_truth = np.array([3.5000000, 2.5000000])
-    y_interp = interp(x_interp, x, y)
-    assert_array_almost_equal(y_interp, y_interp_truth, 7)
-
-
-def test_interp_decrease_xp():
-    """Test interpolation with decreasing order."""
-    x = np.array([4., 3., 2., 1.])
-    y = x
-    x_interp = np.array([3.5000000, 2.5000000])
-    y_interp_truth = np.array([3.5000000, 2.5000000])
-    y_interp = interp(x_interp, x, y)
-    assert_array_almost_equal(y_interp, y_interp_truth, 7)
-
-
-def test_interp_end_point():
-    """Test interpolation with point at data endpoints."""
-    x = np.array([1., 2., 3., 4.])
-    y = x
-    x_interp = np.array([1.0, 4.0])
-    y_interp_truth = np.array([1.0, 4.0])
-    y_interp = interp(x_interp, x, y)
-    assert_array_almost_equal(y_interp, y_interp_truth, 7)
-
-
-def test_interp_masked_units():
-    """Test interpolating with masked arrays with units."""
-    x = np.ma.array([1., 2., 3., 4.]) * units.m
-    y = np.ma.array([50., 60., 70., 80.]) * units.degC
-    x_interp = np.array([250., 350.]) * units.cm
-    y_interp_truth = np.array([65., 75.]) * units.degC
-    y_interp = interp(x_interp, x, y)
-    assert_array_almost_equal(y_interp, y_interp_truth, 7)
 
 
 def test_greater_or_close():
