+++ metpy/interpolate/__init__.py
@@ -5,12 +5,10 @@ r"""Provides tools for interpolating data."""
 
 from .grid import *  # noqa: F403
 from .one_dimension import *  # noqa: F403
-from .physical import *  # noqa: F403
 from .points import *  # noqa: F403
 from .tools import *  # noqa: F403
 
 __all__ = grid.__all__[:]  # pylint: disable=undefined-variable
 __all__.extend(one_dimension.__all__)  # pylint: disable=undefined-variable
-__all__.extend(physical.__all__)  # pylint: disable=undefined-variable
 __all__.extend(points.__all__)  # pylint: disable=undefined-variable
 __all__.extend(tools.__all__)  # pylint: disable=undefined-variable
