+++ metpy/tests/test_gridding.py
@@ -11,6 +11,9 @@ from metpy.deprecation import MetpyDeprecationWarning
 def test_deprecation():
     """Test deprecation warning on import, and that public API functions are available."""
     with pytest.warns(MetpyDeprecationWarning):
-        from metpy.gridding import (interpolate, inverse_distance,  # noqa: F401
-                                    natural_neighbor, remove_nan_observations,
-                                    remove_observations_below_value, remove_repeat_coordinates)
+        import metpy.gridding
+        functions = ['interpolate', 'inverse_distance', 'natural_neighbor',
+                     'remove_nan_observations', 'remove_observations_below_value',
+                     'remove_repeat_coordinates']
+        for function in functions:
+            assert function in metpy.gridding.__all__
