+++ setup.cfg
@@ -17,7 +17,7 @@ flake8-ignore = *.py F405
                 versioneer.py ALL
                 deprecation.py C801
                 metpy/calc/__init__.py D999
-                metpy/gridding/__init__.py D999
+                metpy/interpolate/__init__.py D999
                 metpy/io/__init__.py D999
                 metpy/plots/__init__.py D999
 flake8-max-line-length = 95
